+++ src/iqm/qiskit_iqm/fake_backends/iqm_fake_backend.py
@@ -275,7 +275,7 @@ class IQMFakeBackend(IQMBackendBase):
 
     @classmethod
     def _default_options(cls) -> Options:
-        return Options(shots=1024, calibration_set_id=None)
+        return Options(shots=1024)
 
     @property
     def max_circuits(self) -> Optional[int]:
@@ -315,7 +315,7 @@ class IQMFakeBackend(IQMBackendBase):
             def run(self, dag):
                 qubits_involved_in_last_move = None  # Store which qubit was last used for MOVE IN
                 for node in dag.op_nodes():
-                    if node.op.name not in this.noise_model.basis_gates + ["id", "barrier", "measure", "measurement"]:
+                    if node.op.name not in this.noise_model.basis_gates + ["id", "barrier", "measure"]:
                         raise ValueError("Operation '" + node.op.name + "' is not supported by the backend.")
                     if qubits_involved_in_last_move is not None:
                         # Verify that no single qubit gate is performed on the qubit between MOVE IN and MOVE OUT
@@ -372,8 +372,15 @@ class IQMFakeBackend(IQMBackendBase):
         return job
 
     def validate_compatible_architecture(self, architecture: DynamicQuantumArchitecture) -> bool:
-        """Given a dynamic quantum architecture returns true if its number of components, names of components and
-        component connectivity matches the architecture of this backend."""
+        """Compare a dynamic quantum architecture to the static architecture of the fake backend.
+
+        Args:
+            architecture: dynamic quantum architecture to compare to
+
+        Returns:
+            True iff the number and names of the locus components, the component connectivity,
+            and the available operations in the DQA match the static architecture of this backend.
+        """
         components_match = set(architecture.components) == set(self.__architecture.qubits)
         ops = {
             gate_name: list(list(locus) for locus in gate_info.loci)